fixed bug for fix type

+++ FrskyGps/defines.h
@@ -6,7 +6,7 @@ int        groundspeed = 0;
 float      lat = 0;
 float      lon = 0;
 uint8_t    sats = 0;
-uint16_t   fix = 0;
+int   fix = 0;
 
 byte lat_dir;
 byte lon_dir;
